+++ Lab_Stock/src/Tradable.java
@@ -17,6 +17,6 @@ public interface Tradable {
      */
 
 
-    int getMetric();
+    double getMetric();
 
 }
